requirements.txt
flask
pandas
openpyxl
gunicorn
psycopg2-binary